server/requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
backoff
beautifulsoup4
blinker
cachetools
certifi
cffi
chardet
charset-normalizer
click
colorama
coloredlogs
contourpy
cryptography
cycler
dataclasses-json
deepdiff
emoji
faiss-cpu
fastapi
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
huggingface-hub
humanfriendly
idna
intel-openmp
iopath
Jinja2
joblib
jsonpatch
jsonpath-python
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
langchain
langchain-cli
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-huggingface
langchain-text-splitters
langdetect
langserve
langsmith
layoutparser
libcst
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mkl
mpmath
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
onnx
onnxruntime
opencv-python
ordered-set
orjson
packaging
pandas
pdf2image
pdfminer.six
pdfplumber
pillow
pillow_heif
pip
portalocker
proto-plus
protobuf
psutil
pyarrow
pyasn1
pyasn1_modules
pycparser
pycryptodome
pydantic
pydantic_core
pydeck
Pygments
PyMuPDF
PyMuPDFb
pyparsing
pypdf
PyPDF2
pypdfium2
pyproject-toml
pyreadline3
pytesseract
python-dateutil
python-docx
python-iso639
python-magic
python-magic-bin
python-multipart
python-pptx
pytz
pywin32
PyYAML
rapidfuzz
referencing
regex
requests
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
sse-starlette
starlette
streamlit
sympy
tabulate
tbb
tenacity
threadpoolctl
tiktoken
timm
tokenizers
toml
tomlkit
toolz
torch
torchvision
tornado
tqdm
transformers
typer
typing_extensions
typing-inspect
tzdata
unstructured
unstructured-client
unstructured-inference
unstructured.pytesseract
uritemplate
urllib3
uvicorn
watchdog
wheel
wrapt
XlsxWriter
yarl
